refactor test layout
